requirements.txt
streamlit
torch
Pillow
numpy
opencv-python
sahi
yolov5
streamlit
torch
Pillow
numpy
opencv-python
sahi

